requirements.txt
atproto==0.0.62
beautifulsoup4==4.13.5
deep-translator==1.11.4
fastapi==0.116.1
feedparser==6.0.12
google-genai==1.0.0
google-generativeai==0.8.5
llama-cpp-python[avx2]==0.2.74
pymongo==4.10.0
pillow==10.4.0
pyspellchecker==0.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
regex
requests==2.32.3
safetensors==0.6.2
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
transformers==4.44.2
trafilatura==2.0.0
uvicorn[standard]
yt-dlp